requirements.txt
fastapi
uvicorn
presidio-analyzer
python-multipart
pdfplumber
python-docx
pytesseract
pdf2image
pillow
